+++ private/test/src/helpers.ts
@@ -21,6 +21,7 @@ import { describe, expect } from "vitest";
 
 import type { Client, DecisionLogEntry } from "@cerbos/core";
 import { CheckResourcesResult } from "@cerbos/core";
+import type { BundleMetadata } from "@cerbos/embedded";
 
 import type { QueryServiceClient } from "./protobuf/jaeger/proto/api_v3/query_service";
 import type {
@@ -28,6 +29,9 @@ import type {
   TracesData,
 } from "./protobuf/opentelemetry/proto/trace/v1/trace";
 import { cerbosVersionIsAtLeast } from "./servers";
+
+const { version: embeddedSdkVersion } =
+  require("../../../packages/embedded/package.json") as { version: string };
 
 const { version: grpcSdkVersion } =
   require("../../../packages/grpc/package.json") as { version: string };
@@ -243,17 +247,13 @@ export function describeIfCerbosVersionIsAtLeast(
   return describe.skip;
 }
 
-export function bundleFilePath(
-  commit = "68337848cb3f987627da7381653d82c6d4e368a5",
-): string {
-  return resolve(__dirname, `../bundles/${commit}.wasm`);
-}
+const callIdPattern = /^[0123456789ABCDEFGHJKMNPQRSTVWXYZ]{26}$/;
 
-export const callIdMatcher = expect.stringMatching(
+export const callIdMatcher = expect.stringMatching(callIdPattern);
+
+export const versionDependentCallIdMatcher = expect.stringMatching(
   // cerbosCallId was not available on API responses before 0.33.0
-  cerbosVersionIsAtLeast("0.33.0")
-    ? /^[0123456789ABCDEFGHJKMNPQRSTVWXYZ]{26}$/
-    : /^$/,
+  cerbosVersionIsAtLeast("0.33.0") ? callIdPattern : /^$/,
 );
 
 export async function retry<T>(
@@ -275,6 +275,8 @@ export async function retry<T>(
   }
 }
 
+export const embeddedUserAgent = `cerbos-sdk-javascript-embedded/${embeddedSdkVersion}`;
+
 export const grpcUserAgent = `cerbos-sdk-javascript-grpc/${grpcSdkVersion} grpc-node-js/${grpcJsVersion}`;
 
 export const httpUserAgent = `cerbos-sdk-javascript-http/${httpSdkVersion}`;
@@ -293,3 +295,50 @@ export async function getDecisionLogEntry(
     return entry;
   });
 }
+
+export interface EmbeddedBundle {
+  path: string;
+  etag: string;
+  metadata: BundleMetadata;
+}
+
+function embeddedBundle(
+  etag: string,
+  metadata: BundleMetadata,
+): EmbeddedBundle {
+  return {
+    path: resolve(__dirname, `../bundles/${metadata.commit}.wasm`),
+    etag,
+    metadata,
+  };
+}
+
+export const oldEmbeddedBundle = embeddedBundle(
+  '"9017f17ef4224dc634cf33b25201d463"',
+  {
+    builtAt: new Date(Date.UTC(2024, 2, 25, 11, 45, 14)),
+    commit: "68337848cb3f987627da7381653d82c6d4e368a5",
+    policies: [
+      "cerbos.resource.document.v1",
+      "cerbos.resource.document.v1/test",
+    ],
+    sourceAttributes: {},
+  },
+);
+
+export const newEmbeddedBundle = embeddedBundle(
+  '"89b9908716aa06e61be705156e3ae61c"',
+  {
+    builtAt: new Date(Date.UTC(2025, 1, 10, 9, 44, 24)),
+    commit: "0647f50e09a801132ca15110c6d6ca8a3496f1b0",
+    policies: [],
+    sourceAttributes: {
+      "cerbos.resource.document.v1": {
+        source: "document.yaml",
+      },
+      "cerbos.resource.document.v1/test": {
+        source: "test/document.yaml",
+      },
+    },
+  },
+);